:recycle: Refactor Deep Q-network notebook to work on 8x8 grid
Moving some code blocks around to fix missing variables in functions due to reusing global variables... Adding more print messages and some assertion checks in the functions.

For the model specifically, the epsilon parameter was tweaked a little, to try to get a better result for the 8x8 FrozenLake grid environment. See previous commit for a working 4x4 grid version, as this version is still a work in progress.