:boom: Custom CrevasseCrosserEnv, Double and Dueling DQN on keras-rl
Major shift into using a more OpenAI gym compatible schema for the reinforcement learning environment, so that we can make use of the keras-rl library's built in Double DQN and Dueling DQN out of the box. Work in progress as I haven't found the right hyperparameters! On the (possible) TODO list is to move on to DDPG and A3C if time allows.

Note that some of the custom functions now reside in the CrevasseCrosserEnv class. plot_player_and_map is now in env.render. player_step is now deprecated in favour of the FrozenLake style mechanics.

Model is now more closely based on Mnih et al. 2015's architecture with a 8x8 Conv2D as the first layer, and a fully connected linear output layer. Hyperparameters remain largely the same as previous pure numpy implementation. Keras-rl uses steps instead of episodes, so we try to estimate how many steps there are in n episodes. Training for 30000 steps now, or about 1200 episodes which is down from the 2000 episodes used before. Epsilon is now annealed linearly (read: properly) from 1.0 to 0.1. Memory buffer remains at 1000, and there is a warm up period of 50 steps. Target model now updated every 100 steps instead of every 1 step.