:chart_with_upwards_trend: Upsize DQN input image and brace for hyperopt
Main changes relate to upsizing the Deep Q-Network's input image from a 8x8 image to 128x128! In the Convolutional model, I've swtiched from 'same' padding to 'valid' padding to lower the number of parameters, and care less about the edges since we have such a big image now.

DQN_model is retrained again on 1000 episodes with these new model architecture tweaks. Bonus reward was revised upward, Note also that the way the dqn_model is declared now is through a wrapper function, that is (or will be) compatible with hyperopt down the road.