:construction: Work on rewarding agent for getting closer to an endpoint
Part 2 of enabling targetted reinforcement learning. This commit adds in the end_xy parameter in CrevasseCrosserEnv and contains some experimental code to reward the agent for getting closer to a set endpoint. The reward is incrementally given based on a delta_dist value, noting that the agent is still penalized for taking too many steps.

TODO actually adding this endpoint somehow to the observation... and more tests to make sure that all of this works in practice.