:checkered_flag: Retrain crevasse convnet classifier, reduce checkerboard
Fix up (but not completly removed) some checkerboard artifacts, and fixed a bug on the deconvolution dropout layers that slipped through previously. Had attempted to use bilinear upsampling as in https://distill.pub/2016/deconv-checkerboard/ but could not wrap in properly in keras.

Model was retrained from scratch for 300 epochs on 4 GPUs. No data augmentation was used, because the results surprisingly from augmentation are bad. The dropout is now fixed at 0.1 instead of 0.2*i as before. Final validation F1 score is around 0.7, with precision of 0.75 and recall of 0.74.

Note that the F1 Metric class has some new error handling built in, which works to capture some but not all of the errors caused by NaN outputs. Point I want to make is that F1 Metrics can be used on data augmentation now, but may require 1 epoch without augmentation to bootstrap things?