:boom: Async Advantage Actor-Critic (A3C) model !!!
Keras/Tensorflow implementation of A3C, based heavily on many open source implementations, but I'd like to credit morvanzhou's implementation in particular which is very well structured. This commit cleans out a lot of old cruft from having to hack keras-rl to work.

Reward structure is in a messy state of refactoring. Have tried to simplified it to use deciles for when end_xy is not set. Reward structure for when end_xy is set is still not optimal (read:close to a disaster). Anyone looking at this code now should work on the reward more than anything else.