:children_crossing: Implement curriculum learning and widen vistas
Now that the Reinforcement Learning model sort of works, the observation window is upsized from 128 to 256 pixels to fit in with the crevasse classifier model's output. Have also implemented a form of curriculum learning where distance to the endpoint goal is linearly increased as training progresses.

Training episodes increased from 50000 to 80000. The env.render() function also produces a 4x zoomed in view to more clearly see the player's immediate surroundings.