:construction: A more promising iteration on the tfjs crevasse finder
Retrained the ConvNet to work on RGB image input, and finally release the actual (one-line-ish) of code in crevasse_finder.ipynb that exports the keras model to a tensorflowjs model.json (and artifacts) within the tfjs folder. Update the javascript code to make predictions on 3 band RGB tensors instead of 1 band grayscale tensors. Note also that the AlphaDropout layer was replaced with a standard Dropout (3 commits ago actually) so as to be compatible with tfjs...