:twisted_rightwards_arrows: And then the two worlds collided
Anyhow, this is the magical part where we join the models from the crevasse_finder and route_finder jupyter notebooks. It runs fine, but there's a bug around image and geographic pixel rounding which needs to be fixed before it is production ready.

You'll notice the WMS stuff at the bottom which was the tangent I went along to move things into the javascript world. That sidetrack somehow got me carried away so much that I forgot about publishing this very important piece for a week, my bad!